d/gpsd.init: remove sysV to not jeopardize socket activation
As in many other packages we have to realize that socket activation
(nice to have a reduced set of services running until needed) and
sysV scripts do not work well together.

TL;DR: the invoke.rc start to the sysV script will be mapped to start
the systemd service and therefore all benefit of socket activation is
lost. The days of sysV are over and IMHO we can drop that part of the
past to gain more benefits of the future.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>